Compile only 1st level of scss in the directory
