+++ uui-build/package.json
@@ -15,10 +15,10 @@
     "@csstools/postcss-sass": "^5.0.1",
     "@lerna/package-graph": "6.4.0",
     "@lerna/project": "6.4.0",
-    "@rollup/plugin-commonjs": "24.0.0",
-    "@rollup/plugin-node-resolve": "15.0.1",
+    "@rollup/plugin-commonjs": "25.0.0",
+    "@rollup/plugin-node-resolve": "15.0.2",
     "@rollup/plugin-replace": "5.0.2",
-    "@rollup/plugin-typescript": "11.0.0",
+    "@rollup/plugin-typescript": "11.1.1",
     "@svgr/rollup": "6.5.1",
     "@types/jest": "27.5.1",
     "@types/lodash.defaultsdeep": "4.6.6",
@@ -27,7 +27,7 @@
     "babel-jest": "27.5.1",
     "chalk": "2.4.1",
     "cross-spawn": "7.0.3",
-    "eslint": "8.36.0",
+    "eslint": "8.40.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-config-airbnb-typescript": "17.0.0",
@@ -45,15 +45,15 @@
     "postcss-scss": "^4.0.6",
     "react-app-polyfill": "3.0.0",
     "replace-in-files": "^2.0.3",
-    "rollup": "3.20.2",
+    "rollup": "3.22.0",
     "rollup-plugin-postcss-modules": "2.1.0",
     "rollup-plugin-visualizer": "5.9.0",
     "sass": "^1.62.1",
     "source-map-explorer": "2.5.3",
-    "stylelint": "15.3.0",
+    "stylelint": "15.6.2",
     "stylelint-config-sass-guidelines": "9.0.1",
-    "stylelint-config-standard": "30.0.1",
-    "stylelint-scss": "4.6.0",
+    "stylelint-config-standard": "33.0.0",
+    "stylelint-scss": "5.0.0",
     "svgo": "1.3.2",
     "ts-morph": "17.0.1",
     "ts-node": "10.9.1"
